Implemented MySession class.
The issue was that Flask's default session implementation is limited to 4KB by the browser's
cookie size limit. So I set up a db table, sessions, to hold the info that needs to be tansmitted
from one form page to another. It's not a full session implementation, but seems to serve the
puropse for this app. So far.