Unable to deploy app. Trying std environement.
"Everybody" is having this problem.
Since previous deploy, the std environment provided support for Python 3.7 and
beta support for Postgres, so I'm trying to switch. It eliminates the need for
gunicorn and gevent, and deployment (a) completes successfully and (b) is much
faster than flex where the whole app has to be rebuilt for each deploy (and the)
app never becomes healthy, so the deployment fails.

Unfortunately, the deployed app generates server errrors, with as yet not info
in the logs as to what is going wrong.

Got rid of logging to a file: the logs location is not user-writeable.
Next step is to enable stackdriver debugging to see if the source of the server
error can be located.