Bug fixes and UI enhancement.
One bug was that the Next button was disabled when back-buttoning to the first
page of review rules. (This needs to be looked at for other cases.)
Second bug was finding no rules when not in DEBUG mode. (Indentation error.)
The UI enhancement was to reword the message about number of matching rules
in do_form_2().